New test suite additions; Decimal and Unicode type tests.
Replaced some assert calls that made equality checks with more useful
calls to AssertEqual.

Added code to set the current working directory to the test directory
this allows the test suite to be run from any path as some tests use
data files that are included with the test scripts.

Added warnings about tests that are not ran.

Added new tests:
    test_decimalLongValues - ensure values are selected correctly
    test_decimalBindLongValues - ensure bind values are sent correctly
    test_decimalCrash - crash bug found with (many) decimals
    test_SimpleUnicodeSelect - ensure Unicode data is selected correctly
    NOtest_SimpleUnicodeBind - disabled due to hang, ensure Unicode bind values are sent correctly



git-svn-id: http://code.ingres.com/ingres/drivers/python/main@1865 45b5d43f-8932-4c86-835a-3654e5842839