Fixes IngresDBI Python driver Trac ticket #446 error on execute with empty parm iterable.
It is reasonable to send an empty, or None bind param to cursor.execute, 
added simply check that only attempts to use the bind params if there 
are some.

Created IIDBI_IsTrue(), which is now used to check (bind)
parameters for execute and executemany. For now this is local
to ingresdbi.c but could be renamed/moved if it is found useful.

New test test_FetchAfterClose for Trac ticket 445
New test test_ExecuteWithEmptyParams for Trac ticket 446



git-svn-id: http://code.ingres.com/ingres/drivers/python/main@2085 45b5d43f-8932-4c86-835a-3654e5842839