Fixes Trac ticket #445 Python DBI driver crash under Windows when fetching rows from cursor on closed connection.
After an ODBC connection is closed the ODBC spec requires that the driver close and free all resources. The crash is actually expected!

Added check for exception for to fetchall so that exceptions raised 
by fetchone are not lost.

Added quick/simply fix to check connection status on fetch. The old
approach for the original IngresDBI driver was to maintain a dict of
cursors in the connection and when the connection was closed, close
all the cursors. This is not present in this pure C extension.
Really should return back to that technique. Checking the connection
for each single fetch is a needless overhead.

NOTE problem did not occur under Linux, it appears the Unix Ingres
ODBC driver (CLI manager) returns errors when a cursor for a closed
connection is used (or we just got lucky).


git-svn-id: http://code.ingres.com/ingres/drivers/python/main@2112 45b5d43f-8932-4c86-835a-3654e5842839