Fix for Trac Bug #526 Add libframe library.
Python driver needs ODBC (and CLI),
    which needs OpenAPI,
    which in turn needs libq as that is where existing GCA code is
libq needs libframe, for some reason on some Unix platforms (Solaris)
libframe must be explictly linked in (unlike say Windows and Linux).
However linking in libframe under Linux is not an issue so setting
libframe as a default for all platforms (bar Windows).

This was an issue for other drivers that use OpenAPI (e.g. php and Ruby).
The Ruby driver (which explictly uses OpenAPI) goes as far a linking in:
" -lq.1 -lcompat.1 -lframe.1 -liiapi.1"



git-svn-id: http://code.ingres.com/ingres/drivers/python/main@2456 45b5d43f-8932-4c86-835a-3654e5842839