Fix #161 - Provide a more descriptive error message when a fetch is performed against non-SELECT query.
Overview
--------

When issuing a cursor.fetchall(), cursor.fetchone() or fetchmany() after
executing a non-SELECT SQL statement the ingresdbi extension returns an
InterfaceError exception along with the text "No tuple descriptor count".
Whilst the text is technically correct it was felt that the exception and
message needed to be simplified. This change updates the exception to a
DataError with the text "No results are available". DataError has been
used as it fits better than InterfaceError, since there is no problem
with the actual inferface.

git-svn-id: http://code.ingres.com/ingres/drivers/python/main@3978 45b5d43f-8932-4c86-835a-3654e5842839