Fix trac #702 - tearDown() does not clean up after each test
Overview
--------

When running the test suite using the command:

test_database=demodb /usr/bin/python tests/test_ingresdbi_dbapi20.py -v

a failure in one test can cause a hang in a later test that uses the same
table. This is the result of the connection for the failing test not being
cleaned up, holding locks on the required table.

This change updates the tearDown() method in test_ingresdbi_dbapi20.py so that
after each test is run, any open cursor or connection is closed. This releases
any held locks allowing the test suite to continue to the end.

git-svn-id: http://code.ingres.com/ingres/drivers/python/main@3979 45b5d43f-8932-4c86-835a-3654e5842839