Fix #407 - README.txt/html does not explicity state that ODBC is required
Overview
--------

The README files for the ingresdbi do not explicity state that ODBC is required
to build or use the inferface. This change rectifies this by stating that the
Ingres ODBC driver is a requirement both for building and using.

In addition this change fixes the syntax of the ingresdbi.connect() method as
described in the README. The database attribute was missing an "=" between the
attribute name and the attribute value.

git-svn-id: http://code.ingres.com/ingres/drivers/python/main@3992 45b5d43f-8932-4c86-835a-3654e5842839